Enum mailparse::MailParseError

pub enum MailParseError {
    QuotedPrintableDecodeError(QuotedPrintableError),
    Base64DecodeError(DecodeError),
    EncodingError(Cow<'static, str>),
    Generic(&'static str),
}

An error type that represents the different kinds of errors that may be encountered during message parsing.

Variants
QuotedPrintableDecodeError(QuotedPrintableError)

Data that was specified as being in the quoted-printable transfer-encoding could not be successfully decoded as quoted-printable data.

Base64DecodeError(DecodeError)

Data that was specified as being in the base64 transfer-encoding could not be successfully decoded as base64 data.

EncodingError(Cow<'static, str>)

An error occurred when converting the raw byte data to Rust UTF-8 string format using the charset specified in the message.

Generic(&'static str)

Some other error occurred while parsing the message; the description string provides additional details.