Struct mailparse::ParsedContentType

pub struct ParsedContentType {
    pub mimetype: String,
    pub charset: String,
    pub boundary: Option<String>,
    pub name: Option<String>,
}

A struct to hold a more structured representation of the Content-Type header. This is provided mostly as a convenience since this metadata is usually needed to interpret the message body properly.

Fields
mimetype: String

The type of the data, for example "text/plain" or "application/pdf".

charset: String

The charset used to decode the raw byte data, for example "iso-8859-1" or "utf-8".

boundary: Option<String>

The boundary used to separate the different parts of a multipart message. This boundary is taken straight from the Content-Type header, and so the body will actually contain the boundary string prefixed by two dashes.

name: Option<String>

The name of the content, if available