Struct mailparse::MailHeader

pub struct MailHeader<'a> { /* fields omitted */ }

A struct that represents a single header in the message. It holds slices into the raw byte array passed to parse_mail, and so the lifetime of this struct must be contained within the lifetime of the raw input. There are additional accessor functions on this struct to extract the data as Rust strings.

Methods

impl<'a> MailHeader<'a>

fn get_key(&self) -> Result<String, MailParseError>

Get the name of the header. Note that header names are case-insensitive.

fn get_value(&self) -> Result<String, MailParseError>

Get the value of the header. Any sequences of newlines characters followed by whitespace are collapsed into a single space. In effect, header values wrapped across multiple lines are compacted back into one line, while discarding the extra whitespace required by the MIME format. Additionally, any quoted-printable words in the value are decoded.

Examples

use mailparse::parse_header;
    let (parsed, _) = parse_header(b"Subject: =?iso-8859-1?Q?=A1Hola,_se=F1or!?=").unwrap();
    assert_eq!(parsed.get_key().unwrap(), "Subject");
    assert_eq!(parsed.get_value().unwrap(), "\u{a1}Hola, se\u{f1}or!");